Use var in all dashboard controllers and use to build menus
