+++ module/Olcs/view/correspondence.phtml
@@ -15,9 +15,15 @@ echo $this->partial(
 
       /* @var \Laminas\View\Helper\Navigation\Menu $menu */
       $menu = $this->navigation($this->navigation('navigation')->getContainer()->findBy('id', 'dashboard-licences-applications'))->menu();
-      
-      echo $menu->setMinDepth(0)->setMaxDepth(0)
-      ->setPartial('partials/tabs-nav');
+
+      echo $menu->setMinDepth(0)
+        ->setMaxDepth(0)
+        ->setPartial('partials/tabs-nav')
+        ->renderPartialWithParams(
+          [
+            'unreadMessageCount' => $this->__vars['unreadMessageCount']
+          ]
+        );
 
       if (isset($table)):
         echo $table;
