+++ module/Permits/view/permits/permits/index.phtml
@@ -16,8 +16,13 @@ use Permits\View\Helper\IrhpApplicationSection;
             ->menu();
 
         echo $menu->setMinDepth(0)
-            ->setMaxDepth(0)
-            ->setPartial('partials/tabs-nav');
+          ->setMaxDepth(0)
+          ->setPartial('partials/tabs-nav')
+          ->renderPartialWithParams(
+            [
+              'unreadMessageCount' => $this->__vars['unreadMessageCount']
+            ]
+          );
         ?>
 
         <?php echo $this->flashMessengerAll(); ?>
